backend/requirements.txt
# Python dependencies for RDP WebSocket proxy

# WebSocket server
websockets>=12.0

# Environment variables
python-dotenv>=1.0.0

# Image processing for frame compression
Pillow>=10.0.0

# Optional: async HTTP client (for future REST API)
httpx>=0.25.0
